Update quiz route and quiz mock data
