+++ components/quiz/Answer.tsx
@@ -1,4 +1,4 @@
-const Answer = () => {
+const Answer = (): JSX.Element => {
   return <div>test</div>
 }
 export default Answer
