+++ package.json
@@ -10,6 +10,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "nanoid": "^3.1.23",
     "next": "^10.2.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
