+++ pages/index.tsx
@@ -18,7 +18,7 @@ export default function Home(): JSX.Element {
 
         <div className="container flex flex-col mx-auto w-full items-center justify-center">
           <ul className="flex flex-col">
-            <Link href="/quiz_all">
+            <Link href="/quiz">
               <li className="border-gray-400 flex flex-row mb-2">
                 <div className="shadow border select-none cursor-pointer bg-white dark:bg-gray-800 rounded-md flex flex-1 items-center p-4">
                   <div className="flex flex-col w-10 h-10 justify-center items-center mr-4">
