+++ public/data/QuestionBank.json
@@ -1,0 +1,29 @@
+[
+    {
+        "id":"hB3qZ_f_CI0EsD_WRsRX8",
+        "modules": ["CS2030", "CS2030S"],
+        "type": "MRQ",
+        "difficulty": "easy", 
+        "question": "What are valid format specifiers for the String.format() method?",
+        "correct_answers": ["%f", "%d","%s"], 
+        "incorrect_answers": ["%w","%z","%q"]
+    },
+    {
+        "id":"CxcBhgLQeq_ez8MxeDe0p",
+        "modules": ["CS2030", "CS2030S"],
+        "type": "MRQ",
+        "difficulty": "easy", 
+        "question": "In Java, what does the method signature consist of? Choose all the applicable options.",
+        "correct_answers": ["Method name", "Parameter list (number, type and order of its parameters)"], 
+        "incorrect_answers": ["Access modifier","Return type","Exception", "Method body"]
+    },
+    {
+        "id":"dgC_OXQwPq5MG_VwMQLLz",
+        "modules": ["CS2030", "CS2030S"],
+        "type": "MCQ",
+        "difficulty": "easy", 
+        "question": "What is a package in java?",
+        "correct_answers": ["Collection of related classes and interfaces which are bundled together"], 
+        "incorrect_answers": ["Collection of related classes which are bundled together","Collection of related interfaces which are bundled together","Collection of related objects and classes which are bundled together", "None of the others"]
+    }
+]
