+++ public/data/QuizData.json
@@ -1,26 +1,26 @@
 [
     {
-        "modules": ["CS2030", "CS2030S"],
-        "type": "MRQ",
-        "difficulty": "easy", 
-        "question": "What are valid format specifiers for the String.format() method?",
-        "correct_answers": ["%f", "%d","%s"], 
-        "incorrect_answers": ["%w","%z","%q"]
+        "id":"D0YlLZVQFm21OvnDH-9yX",
+        "title":"CS2030 Quiz 1",
+        "author":"Yongliang",
+        "questions": ["hB3qZ_f_CI0EsD_WRsRX8", "CxcBhgLQeq_ez8MxeDe0p","dgC_OXQwPq5MG_VwMQLLz"]
     },
     {
-        "modules": ["CS2030", "CS2030S"],
-        "type": "MRQ",
-        "difficulty": "easy", 
-        "question": "In Java, what does the method signature consist of? Choose all the applicable options.",
-        "correct_answers": ["Method name", "Parameter list (number, type and order of its parameters)"], 
-        "incorrect_answers": ["Access modifier","Return type","Exception", "Method body"]
+        "id":"TxRDDtSEpnpodO0d-3ygJ",
+        "title":"CS2030 Quiz 2",
+        "author":"Yongliang",
+        "questions": ["hB3qZ_f_CI0EsD_WRsRX8", "CxcBhgLQeq_ez8MxeDe0p","dgC_OXQwPq5MG_VwMQLLz"]
     },
     {
-        "modules": ["CS2030", "CS2030S"],
-        "type": "MCQ",
-        "difficulty": "easy", 
-        "question": "What is a package in java?",
-        "correct_answers": ["Collection of related classes and interfaces which are bundled together"], 
-        "incorrect_answers": ["Collection of related classes which are bundled together","Collection of related interfaces which are bundled together","Collection of related objects and classes which are bundled together", "None of the others"]
+        "id":"DnFFpNWyXk0gIwMlxP4Z2",
+        "title":"CS2030 Quiz 3",
+        "author":"Yongliang",
+        "questions": ["hB3qZ_f_CI0EsD_WRsRX8", "CxcBhgLQeq_ez8MxeDe0p","dgC_OXQwPq5MG_VwMQLLz"]
+    },
+    {
+        "id":"2MCYftI-pKIeQ3fOuOpOw",
+        "title":"CS2030 Quiz 4",
+        "author":"Yongliang",
+        "questions": ["hB3qZ_f_CI0EsD_WRsRX8", "CxcBhgLQeq_ez8MxeDe0p","dgC_OXQwPq5MG_VwMQLLz"]
     }
 ]
